+++ lib/generators/simple_form/templates/config/initializers/simple_form.rb
@@ -118,6 +118,10 @@ SimpleForm.setup do |config|
   # matches the regexp as value.
   # config.input_mappings = { /count/ => :integer }
 
+  # Custom wrappers for input types. This should be a hash containing an input
+  # type as key and the wrapper that will be used for all inputs with specified type.
+  # config.wrapper_mappings = { :string => :prepend }
+
   # Default priority for time_zone inputs.
   # config.time_zone_priority = nil